shared-records-report: move to shared code
